Change close-window to use 'a' instead
Much like closing a tab in a browser is C-w instead of C-d, perhaps
closing a thing should be inconsistent depending on what the thing is,
just so that it is a very intentional action. Pressing super+ctrl+a is
a very natural motion when ctrl is mapped to capslock
(thumb+pinky+ring), so it serves this frequent action well while being
unique enough so as to inspire thought and prevent accidental or
muscle-memory-induced window closing.